+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Update `exp`, `mod`, and `sqrt` functions' input parameter type from `float64` to `number`.
 - Update `sqrt` return type to `number`.
 - Verify `round` return does not under/overflow.
+- Permit `number_of_characters` to be equal to `string` parameter length.
 
 ### 2.2.1
 - Dynamic input emptiness error message now logs accurate parameter position.